refactor: move ir-related files to subfolder
