+++ src/app-data/kinds/ir/__tests__/getIRMeasurement.test.ts
