+++ src/app-data/kinds/ir/__tests__/irMeasurementEnhancer.test.ts
@@ -1,8 +1,9 @@
 import { produce } from 'immer';
 import { expect, test } from 'vitest';
 
-import { getIRMeasurement } from '../../__tests__/getIRMeasurement';
 import { irMeasurementEnhancer } from '../irMeasurementEnhancer';
+
+import { getIRMeasurement } from './getIRMeasurement';
 
 test('irMeasurementEnhancer', async () => {
   const measurement = await getIRMeasurement();
