+++ src/app-data/kinds/ir/irMeasurementEnhancer.ts
@@ -1,4 +1,4 @@
-import type { MeasurementBase } from '../MeasurementBase';
+import type { MeasurementBase } from '../../MeasurementBase';
 
 /**
  * we add missing absorbance / transmittance
